Ready to do some actual work
Was messing around with the fucking display.

It's nonsense. I should be finishing the fucking PPU, not messing around with how to display fucking nametables.

Nametables should only be displayed as a debug option. And I don't think it makes any sense to debug
something that hasn't actually happened yet... Dumb. Just dumb