Performance is not. Good.
We might actually need to parallelize CPU and PPU. It would actually be kind of cool.
Since RAM is a mutable vector, they can very easily share data.

The most difficult problems are

1. CPU can interact with the PPU via addresses $2000 to $2007. To solve this, we would use two TChan: ptc and ctp.

CPU would write to ctp with a request. (Read or Write | Address | [Byte]) and so on.
It would then read on ptc and hang until it receives a response (the byte read, or an OK, idk).

The PPU would maybeRead on ctp. On successfull reads, it would send the information required via ptc.

2. Syncronization. The PPU runs at roughly 3 times the PPU speed. In order to make sure they match, we would need to do

a) Artificially slow down the CPU so it is roughly 3 times slower than the PPU.
b) Use a TChan to make the CPU wait until 3 iterations of the PPU have finished. I don't know if this would actually have a positive effect on performance.

Idk. Perhaps these are all shite ideas though. Fuck it. Idk